requirements.txt
azure-functions
azure-eventhub
requests
python-dotenv
pytest